agent-driven: Move FieldNumericOverrides to numeric_override module
